csdf/requirements.txt
opencv-python>=4.5.5
numpy>=1.23
Pillow>=9.0
flask>=2.0
matplotlib>=3.5
tqdm>=4.60
scipy>=1.8
